backend/requirements.txt
fastapi
uvicorn
python-multipart
sentence-transformers
torch
whisper
numpy
spacy
numpy
Pillow
mss
easyocr
pytesseract
opencv-python
pywin32
pywinauto
psutil
torch
torchvision
ultralytics
tf-keras
kokoro
soundfile